Enhance Perlin noise module and add new features
- Updated documentation for the Perlin noise module to provide clearer usage examples and descriptions.
- Improved exception handling in the PerlinNoise class.
- Added support for 1D, 2D, and 3D noise generation with separate methods for each.
- Introduced octave noise generation methods for more complex noise patterns.
- Implemented a generate_noise_map function for creating 2D noise maps.
- Added utility functions for numpy integration and terrain generation.
- Included hardware acceleration information for OpenCL support.
- Added version information to the module.

Refactored SHA1, Snowflake, and TEA modules for consistency and clarity

- Removed unnecessary whitespace in SHA1 module.
- Updated Snowflake module to use mutex for thread safety.
- Included random header in TEA module for better random number generation.

Introduced new fnmatch module for enhanced pattern matching

- Implemented fnmatch module with features like case-insensitive matching, path-aware matching, and SIMD acceleration.
- Added support for multiple patterns and parallel processing options.
- Provided comprehensive documentation and examples for usage.

Added MD5 hashing module with binary support

- Created MD5 module with optimized hashing functions for strings and binary data.
- Included exception handling for encryption failures.
- Provided detailed documentation and examples for MD5 usage.

+++ python/algorithm/matrix_compress.cpp
@@ -4,7 +4,6 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-
 namespace py = pybind11;
 
 PYBIND11_MODULE(matrix_compress, m) {
@@ -419,8 +418,8 @@ Examples:
             Matrix matrix(buf.shape[0], std::vector<char>(buf.shape[1]));
             char* ptr = static_cast<char*>(buf.ptr);
 
-            for (size_t i = 0; i < buf.shape[0]; i++) {
-                for (size_t j = 0; j < buf.shape[1]; j++) {
+            for (size_t i = 0; i < static_cast<size_t>(buf.shape[0]); i++) {
+                for (size_t j = 0; j < static_cast<size_t>(buf.shape[1]); j++) {
                     matrix[i][j] = ptr[i * buf.shape[1] + j];
                 }
             }
